Rearrange the board and add a color selecting swatch to the side
This places the resize board button to the side and creates a color picking
swatch below that button. It shoudl also dynamically resize the board based
on the screen width an height so the board shouldn't go to far out of the
screen. Clicking on one of the color swatches to the side switches the color
of the etching. I also added a boarder around the etch area.