Add fade In, Fade out, and randome color buttons and modes
Adds the buttons to change the opacity of the current color you can fade
in which increases the opacity up to 1. Fade out which decreases the opacity
down to 0, or color a tile randomly which generates rgb(0-255,0-255,0-255)
values instead of the selected color. a pointer image indicating the
selected color was added. Comments were added to explain the java script.